First attempt to do this
This is NOT perfect yet. What I want is that if you start casting while
the modifier would expire before you finish casting, the modifier would
extend once and only once. Now it does that, but interrupting when you
have less than 0.25s (cast point) left of the modifier the modifier is
just destroyed.